генерация вопроса вынесена в отдельный файл
