+++ src/mixin/questionGeneratorMixin.ts
@@ -1,0 +1,11 @@
+import { defineComponent } from 'vue'
+
+import QuestionGenerator from '@/QuestionGenerator'
+
+export default defineComponent({
+  data() {
+    return {
+      questionGenerator: new QuestionGenerator()
+    }
+  }
+})
